requirements.txt
cupy_cuda12x==13.1.0
kornia==0.7.2
matplotlib==3.8.4
NLSE==2.3.0
numpy==2.0.0
scipy==1.13.1
scikit-image==0.22.0
scikit-learn==1.3.0
torch==2.3.0
torchvision==0.15.2
tqdm==4.66.1